macro support
By default shuttle open’s new commands in a new tab. These changes
allow you to set flag “isNewTab”: “false” if set the command will open
in the active tab. This is useful if you want to run a commonly used
command in the existing ssh tab.

  {
     "name": "check shuttle process",
      "isNewTab": "false",
      "cmd": "ps aux | grep [s]sh”
  }

With isNewTab set to false the command will run in your existing
terminal tab and not a new one.

I branched these changes off my terminal-themes as some of the code was
reused.