specify custom path to the json file
The ability to store the json file in a custom path is useful
especially if your using the same json file on several systems through
a service like dropbox, one drive, box etc…
These changes allow you set a hard coded path to the *.json file. Do to
this you would create a new file in your home directory called
.shuttle.path so ~/.shuttle.path would be the name and location.
In this file is the path to the json file. Mine currently reads
/Users/thshdw/Desktop/shuttle.json

if the shuttle.path file does not exist then the system reads the file
in its default location ~/shuttle.json. This ensures that this version
is backwards compatible with existing installs.