iTerm Nightly support - Run commands in new or existing windows - Set custom path
iTerm support for stable and nightly versions. Set “stable” or
“nightly” in the JSON settings. Hardcoded applescript has been removed.
In testing compiled applescripts were more consistent. Applescripts are
configured with a handler and bundled with the app.

Commands can be run in a new window or in the current window. Set
“inTerminal” to “new” or “current”. Current window will always be the
front active window.

You can set a custom path for the JSON files. Do to this you would
create a new file in your home directory called .shuttle.path

~/.shuttle.path

In this file is the path to the JSON settings. Mine currently reads
/Users/thshdw/Desktop/shuttle.json

If the shuttle.path file does not exist then the system reads the file
in its default location ~/shuttle.json. This ensures that this version
is backwards compatible with existing installs.